Removed unused classes.
