--- src/io/CSVParserException.java
@@ -1,14 +1,0 @@
-package io;
-
-/**
- * The type Csv parser exception.
- */
-public class CSVParserException extends Exception{
-    /**
-     * Instantiates a new Csv parser exception.
-     *
-     */
-    CSVParserException() {
-        super("There was a fatal error");
-    }
-}
